chore: move BridgeViewMode into types
