+++ app/components/UI/Bridge/hooks/useInitialDestToken.ts
@@ -4,14 +4,15 @@ import { useEffect } from 'react';
 import { DefaultSwapDestTokens } from '../constants/default-swap-dest-tokens';
 import { selectChainId } from '../../../../selectors/networkController';
 import { RouteProp , useRoute } from '@react-navigation/native';
-import { BridgeRouteParams , BridgeViewMode } from './useInitialSourceToken';
+import { BridgeRouteParams } from './useInitialSourceToken';
+import { BridgeViewMode } from '../types';
 
 export const useInitialDestToken = () => {
   const route = useRoute<RouteProp<{ params: BridgeRouteParams }, 'params'>>();
   const dispatch = useDispatch();
   const chainId = useSelector(selectChainId);
 
-  const isSwap = route.params.title === BridgeViewMode.Swaps;
+  const isSwap = route.params.title === BridgeViewMode.Swap;
 
   useEffect(() => {
     const defaultDestToken = DefaultSwapDestTokens[chainId];
