+++ app/components/UI/Bridge/hooks/useInitialSourceToken.ts
@@ -2,7 +2,7 @@ import { useRoute, RouteProp } from '@react-navigation/native';
 import { setSourceToken } from '../../../../core/redux/slices/bridge';
 import { useDispatch, useSelector } from 'react-redux';
 import { useEffect } from 'react';
-import { BridgeToken } from '../types';
+import { BridgeToken, BridgeViewMode } from '../types';
 import { selectEvmNetworkConfigurationsByChainId } from '../../../../selectors/networkController';
 import { useSwitchNetworks } from '../../../Views/NetworkSelector/useSwitchNetworks';
 import { useNetworkInfo } from '../../../../selectors/selectedNetworkController';
@@ -27,12 +27,6 @@ const getNativeSourceToken = (chainId: Hex | CaipChainId) => {
 
   return nativeSourceTokenFormatted;
 };
-
-// TODO: remove this once we move to Unified Swaps
-export enum BridgeViewMode {
-  Swaps = 'Swaps',
-  Bridge = 'Bridge',
-}
 
 export interface BridgeRouteParams {
   token: BridgeToken;
